+++ src/isa/rv64i/environment.rs
@@ -183,7 +183,7 @@ pub const RV64I_SET_E: [Instruction; 2] = [
 
                     match res {
                         Ok(len) => {
-                            cpu.write_x_u64(ABIRegister::A(0).to_x_reg_id() as u8, len)?;
+                            cpu.write_x_u64(ABIRegister::A(0).to_x_reg_id() as u8, len as u64)?;
                         }
                         Err(_e) => {
                             cpu.write_x_i32(ABIRegister::A(0).to_x_reg_id() as u8, -1)?;
@@ -212,7 +212,7 @@ pub const RV64I_SET_E: [Instruction; 2] = [
                     let addr = cpu.read_x_u64(ABIRegister::A(0).to_x_reg_id() as u8)?;
                     cpu.debug_print(|| format!("brk call: {:#x}", addr));
                     if addr != 0 {
-                        cpu.program_brk = addr as u32; // TODO: 64 bit
+                        cpu.program_brk = addr as u64;
                     }
                     cpu.write_x_u64(
                         ABIRegister::A(0).to_x_reg_id() as u8,
