+++ src/isa/rv64i/load_store.rs
@@ -20,7 +20,7 @@ pub const RV64I_SET_LS: [Instruction; 8] = [
                 .wrapping_add_signed(extended_offset);
             let read_value = u8_to_i8(cpu.read_mem_u8(moved_addr)?) as i32;
 
-            cpu.write_x_u64(instruction.rd.value(), read_value as u32)?;
+            cpu.write_x_u64(instruction.rd.value(), read_value as u64)?;
 
             Ok(())
         },
@@ -39,7 +39,7 @@ pub const RV64I_SET_LS: [Instruction; 8] = [
                 .wrapping_add_signed(extended_offset);
             let read_value = u16_to_i16(cpu.read_mem_u16(moved_addr)?) as i32;
 
-            cpu.write_x_u64(instruction.rd.value(), read_value as u32)?;
+            cpu.write_x_u64(instruction.rd.value(), read_value as u64)?;
 
             Ok(())
         },
@@ -72,7 +72,7 @@ pub const RV64I_SET_LS: [Instruction; 8] = [
 
             cpu.debug_print(|| format!("LW: {:#x}", read_value));
 
-            cpu.write_x_u64(instruction.rd.value(), read_value)?;
+            cpu.write_x_u64(instruction.rd.value(), read_value as u64)?;
 
             Ok(())
         },
@@ -91,7 +91,7 @@ pub const RV64I_SET_LS: [Instruction; 8] = [
                 .wrapping_add_signed(extended_offset);
             let read_value = cpu.read_mem_u8(moved_addr)?;
 
-            cpu.write_x_u64(instruction.rd.value(), read_value as u32)?;
+            cpu.write_x_u64(instruction.rd.value(), read_value as u64)?;
 
             Ok(())
         },
@@ -110,7 +110,7 @@ pub const RV64I_SET_LS: [Instruction; 8] = [
                 .wrapping_add_signed(extended_offset);
             let read_value = cpu.read_mem_u16(moved_addr)?;
 
-            cpu.write_x_u64(instruction.rd.value(), read_value as u32)?;
+            cpu.write_x_u64(instruction.rd.value(), read_value as u64)?;
 
             Ok(())
         },
@@ -137,7 +137,7 @@ pub const RV64I_SET_LS: [Instruction; 8] = [
                 )
             });
 
-            cpu.write_mem_u32(moved_addr, read_value)?;
+            cpu.write_mem_u32(moved_addr, read_value as u32)?;
 
             Ok(())
         },
